requirements.txt
django==4.2
djangorestframework
django-cors-headers
drf-yasg
Pillow
django-environ